Trying to optimize SWEP:GetProcessedValue()
The main idea of "optimization" is to localize most of the calls. This allows us to avoid unnecessary calls to "__index"

A LITTLE performance could be given by working with conditions, but I'm not sure how this function works, so I didn't experiment too much - I changed the order of conditions for `UBGL`, `EmptyUBGL`, `Silenced`. See comment "-- !! changed the order".